# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Grok2API
  version: 1.0.0-go
  description: |
    Grok CLI credential gateway. Inference endpoints use OpenAI/Anthropic-shaped
    errors. Admin endpoints use the `{ok,data,error}` envelope.
servers:
  - url: /
...
paths:
...
  /v1/models:
    get:
      tags: [inference]
      summary: List models
      operationId: listModels
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      responses:
        "200":
          description: OpenAI-style model list
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
  /v1/billing:
    get:
      tags: [inference]
      summary: Billing probe
      operationId: getBilling
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      responses:
        "200":
          description: Billing payload
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
  /v1/chat/completions:
    post:
      tags: [inference]
      summary: Chat Completions
      operationId: createChatCompletion
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Chat completion or SSE stream
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "503":
          $ref: "#/components/responses/OpenAIUnavailable"
  /chat/completions:
    post:
      tags: [inference]
      summary: Chat Completions alias
      operationId: createChatCompletionAlias
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      responses:
        "200":
          description: Chat completion or SSE stream
  /v1/responses:
    post:
      tags: [inference]
      summary: OpenAI Responses
      operationId: createResponse
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Responses payload or SSE stream
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
  /v1/messages:
    post:
      tags: [inference]
      summary: Anthropic Messages
      operationId: createMessage
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Messages payload or SSE stream
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
...
  /api/admin/v1/accounts/import/preview:
    post:
      tags: [admin]
      summary: Import accounts (dry-run)
      operationId: adminImportPreview
      security:
        - AdminBearer: []
      responses:
        "200":
          description: Preview result
  /api/admin/v1/accounts/import:
    post:
      tags: [admin]
      summary: Import accounts
      operationId: adminImport
      security:
        - AdminBearer: []
      responses:
        "200":
          description: Import result
...
components:
  securitySchemes:
    ClientBearer:
      type: http
      scheme: bearer
      description: Persisted client API key
    ClientAPIKey:
      type: apiKey
      in: header
      name: x-api-key
    AdminBearer:
      type: http
      scheme: bearer
      description: Short-lived administrator access token
...
  responses:
    OpenAIUnauthorized:
      description: Invalid client API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIError"
    OpenAIRateLimited:
      description: Rate limited / quota circuit
      headers:
        Retry-After:
          schema: { type: string }
        X-Grok2API-Pool-Reason:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIError"
    OpenAIUnavailable:
      description: Pool capacity unavailable
      headers:
        Retry-After:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIError"
...
  schemas:
...
    OpenAIError:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string }
            code: { type: string }
            param: { nullable: true }
...